Fix remaining i64/u64 shims: resolve errors, Syscall enum, translate()
Final pass from audit — change the last few i64/u64 to isize/usize:

- resolve_with_cwd/resolve_parent_and_name: error type i64 → isize
  (errno values are small negative integers, not file offsets)
- Syscall::ArchSpecific/Unknown variants: u64 → usize
- SyscallTranslate trait: nr u64 → usize
- translate_x86_64/translate_aarch64: nr u64 → usize

Every remaining u64/i64 in the codebase is now genuinely fixed-width:
hardware registers, inode IDs, file offsets, ELF64 fields, physical
addresses, timestamps, PTE flags, or Linux ABI struct fields.

+++ crates/kernel/src/arch/aarch64/syscall.rs
@@ -33,7 +33,7 @@ pub fn handle_syscall(frame: *mut u8) {
             }
             221 => { crate::syscall::generic_exec::<super::Aarch64>(a0 as usize, a1 as usize); 0 }
             _ => {
-                let sc = translate_aarch64(nr);
+                let sc = translate_aarch64(nr as usize);
                 crate::syscall::dispatch(sc, a0 as usize, a1 as usize, a2 as usize, a3 as usize, a4 as usize) as i64
             }
         };
@@ -44,7 +44,7 @@ pub fn handle_syscall(frame: *mut u8) {
 }
 
 /// aarch64 Linux syscall number → generic Syscall enum.
-fn translate_aarch64(nr: u64) -> crate::syscall::Syscall {
+fn translate_aarch64(nr: usize) -> crate::syscall::Syscall {
     use crate::syscall::Syscall;
     match nr {
         // File I/O
